# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  library-management-system-go_internal_dto.CreateArticleCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  library-management-system-go_internal_dto.CreateArticleRatingRequest:
    properties:
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
  library-management-system-go_internal_dto.CreateArticleRequest:
    properties:
      category:
        type: string
      content:
        type: string
      excerpt:
        type: string
      featured_image_url:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - category
    - content
    - title
    type: object
  library-management-system-go_internal_dto.CreateArticleReviewRequest:
    properties:
      article_id:
        type: integer
      feedback:
        type: string
      status:
        type: string
    required:
    - article_id
    - status
    type: object
...
  library-management-system-go_internal_dto.UpdateArticleCommentRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  library-management-system-go_internal_dto.UpdateArticleRatingRequest:
    properties:
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
  library-management-system-go_internal_dto.UpdateArticleRequest:
    properties:
      category:
        type: string
      content:
        type: string
      excerpt:
        type: string
      featured_image_url:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  library-management-system-go_internal_dto.UpdateArticleReviewRequest:
    properties:
      feedback:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for school library management.
  title: School Library Management API
  version: "1.0"
paths:
...
  /articles:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      parameters:
      - description: Create article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - articles
  /articles/{id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete article
      tags:
      - articles
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get article
      tags:
      - articles
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - articles
  /articles/{id}/comments:
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List article comments
      tags:
      - article-comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create article comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article comment
      tags:
      - article-comments
  /articles/{id}/comments/{comment_id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete article comment
      tags:
      - article-comments
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: integer
      - description: Update article comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article comment
      tags:
      - article-comments
  /articles/{id}/ratings:
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List article ratings
      tags:
      - article-ratings
    post:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create article rating request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleRatingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article rating
      tags:
      - article-ratings
  /articles/{id}/ratings/{rating_id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete article rating
      tags:
      - article-ratings
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rating ID
        in: path
        name: rating_id
        required: true
        type: integer
      - description: Update article rating request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article rating
      tags:
      - article-ratings
  /articles/review:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List article reviews
      tags:
      - article-reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Create article review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateArticleReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create article review
      tags:
      - article-reviews
  /articles/review/{id}:
    get:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get article review
      tags:
      - article-reviews
    put:
      consumes:
      - application/json
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update article review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateArticleReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update article review
      tags:
      - article-reviews
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"